% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add}
\alias{add}
\title{Sum of two numbers}
\usage{
add(x, y, z)
}
\arguments{
\item{x}{single numeric}

\item{y}{single numeric}

\item{z}{single numeric}
}
\value{
the sum of x, y, and z.
}
\description{
\code{add} returns the sum of x, y, and z.
}
\examples{
 sudokusolvr:::add(3, 4, 6)

}
