% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multi}
\alias{multi}
\title{Product of two elements.}
\usage{
multi(x, y)
}
\arguments{
\item{x}{single numeric}

\item{y}{single numeric}
}
\value{
the product of x and y.
}
\description{
\code{multi} returns the product of two arguments x and y.
}
\examples{
 multi(10, 10)
 multi(5, 5)

}
