% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puzzle1}
\alias{puzzle1}
\title{A 9x9 matrix representation of a sudoku puzzle}
\format{a matrix with 9 rows and 9 columns}
\usage{
puzzle1
}
\description{
a matrix object with 9 rows and 9 columns which is a representation of a
sudoku puzzle. Cells in the soduku puzzle which do not contain a number are
filled with \code{NA}.
}
\keyword{datasets}
