% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puzzle1Array}
\alias{puzzle1Array}
\title{A 9x9x9 array representing solutions of a sudoku puzzle}
\format{a 9x9x9 array}
\usage{
puzzle1Array
}
\description{
a logical array object where the first and second dimensions represent the
rows and columns of a sudoku puzzle and the third dimension represnts which
numbers 1-9 can go in a particular row/column combination. Cells in the array
in which it is unknown whether that number can go in the row/column
combination are filled with \code{NA}
}
\keyword{datasets}
