% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver}
\alias{solver}
\title{Function which solves a sudoku puzzle}
\usage{
solver(puz)
}
\arguments{
\item{puz}{9x9 matrix representation of a sudoku puzzle. Cells of the sudoku
puzzle which are not filled in should be represented by \code{NA} values.}
}
\value{
Takes as input a 9x9 representation of a sudoku puzzle and finds and
  returns its solution. This function utilizes the functions
  \code{initArray}, \code{logic1}, \code{logic2}, \code{upArray} recursively
  in order to find the solution for the puzzle
}
\description{
Takes a 9x9 matrix that represents a sudoku puzzle and returns the puzzle
solved.
}
