% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{sudokuPuzzle}
\alias{sudokuPuzzle}
\title{Creates a sudokuPuzzle class object}
\usage{
sudokuPuzzle(x)
}
\description{
The function \code{sudokuPuzzle()} creates a sudoku puzzle object, which is a
list that contains the sudoku puzzle as the first list item and a 9x9x9
logical array which is used to find and represent the solutions to the puzzle
}
