% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{upArray}
\alias{upArray}
\title{Update puzzle array given values i, j, and k}
\usage{
upArray(puzArr, i, j, k)
}
\arguments{
\item{puzArr}{9x9x9 array corresponding to the sudoku puzzle to be solved.}

\item{i, j}{integer between 1 and 9 referrencing the i x jth element of the
sudoku puzzle matrix and puzzle array}

\item{k}{integer between 1 and 9. Represents the value that goes in the i x
jth cell of the puzzle matrix and is the third dimension index of the
puzzle array.}
}
\value{
Returns an updated puzzle array. Given that k goes in the i x jth
  element of the puzzle matrix, the puzzle array is updated accordingly by
  placing \code{TRUE}/\code{FALSE} values in the appropriate places in the
  puzzle array.

  Function is used within \code{initArray} to initialize the puzzle array
  given the starting values of the sudoku puzzle.
}
\description{
This function takes a puzzle array and a solution for the puzzle given as an
\code{[i, j, k]} index and update the array given the provided solution.
}
\examples{
\dontrun{upArray(b, 1, 1, 1)}

}
