% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{upPuz}
\alias{upPuz}
\title{Function to update puzzle and puzzle array when solution is found}
\usage{
upPuz(puzList, solution)
}
\arguments{
\item{puzList}{is a list object which contains the 9x9 matrix representation
of the sudoku puzzle as the first list element and the 9x9x9 puzzle array
representing the solutions of the sudoku puzzle}

\item{solution}{this is a vector of three numerics which are the index of the
solution found}
}
\value{
This function takes a puzzle list object and a solution produced from
  \code{logic1} and updates the 9x9 puzzle matrix with the corresponding
  solution and places the correct \code{TRUE/FALSE} values in the
  corresponding places in the 9x9x9 puzzle solution array
}
\description{
Takes as input the puzzle list and a solution produced from the \code{logic1}
function
}
